requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
httpx==0.25.2
Pillow==10.1.0
qrcode==7.4.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
flask==3.0.0
gunicorn==21.2.0
pillow==10.1.0
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2
cryptography==41.0.7
bcrypt==4.1.2
schedule==1.2.0
asyncpg==0.29.0
aiosqlite==0.19.0
psutil==5.9.6
pycryptodome==3.20.0
qrcode[pil]==7.4.2
phonenumbers==8.13.26
email-validator==2.1.0
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-compress==1.14
flask-caching==2.1.0
stripe==7.9.0
mercadopago==2.2.1
gerencianet==3.0.0
